+++ trello-parser.py
@@ -15,6 +15,9 @@ paramsField = (
     ('token', token)
 )
 
+params = (('key', key),
+          ('token', token))
+
 
 # Get content of all Column Cards in JSON Format for a given Column
 def accessColumnCards(columnId, params=paramsField):
@@ -25,7 +28,7 @@ def accessColumnCards(columnId, params=paramsField):
 
 
 # Filter Cards of a Column and only return Cards in which the given fieldId was set
-def filterCardsByCustomerField(fieldId, cardsJson):
+def filterCardsByCustomField(fieldId, cardsJson):
     filteredCardsDic = {}
     for i in cardsJson:
         name = i['name']
@@ -59,6 +62,22 @@ def accessCustomFields(boardId):
                     customFieldValueDic[ids] = value
 
     return customFieldDic, customFieldValueDic
+
+
+# Get a dictionary hosting all Label Id and values for a given board
+def accessBoardLabels(boardId):
+    boardUrl = ''.join(['https://api.trello.com/1/boards/', boardId, '/labels'])
+    labelResponse = requests.get(boardUrl, params=params)
+    labelJson = labelResponse.json()
+    print(labelJson)
+    labelDic = {}
+
+    for i in labelJson:
+        name = i['name']
+        ids = i['id']
+        labelDic[name] = ids
+
+    return labelDic
 
 
 # Access Column and retrieve the information stored on a card and in its CustomFields
@@ -98,6 +117,53 @@ def getColumnWithAllFields(boardId, columnId):
     return dataFrame, columnDic
 
 
+# Generates a dataframe that differentiates the story points for done tasks(labeled with approved in the review
+# column) and unfinished tasks
+def generateStoryPointsDataFrame(boardId, taskColumnsIds, finishedTaskColumn, storyPointFieldId):
+    # Generates Dictionary Hosting the Label id and their corresponding value
+    labelDic = accessBoardLabels(boardId)
+
+    # Gather all Tasks, that are labeled with 'approved' in the review column
+    finishedTasksCards = accessColumnCards(finishedTaskColumn, params)
+
+    finishedTasks = []
+    for i in finishedTasksCards:
+        currentLabel = i['idLabels']
+        approvedLabel = labelDic['approved']
+        if approvedLabel in currentLabel:
+            finishedTasks.append(i['name'])
+
+    # Since we want to know how many story points are defined in the current sprint we need to add up the defined
+    # values of all 3 columns
+    storyPointsDic = {}
+
+    for i in taskColumnsIds:
+        currentColumn = accessColumnCards(i)
+        foundTasksDic = filterCardsByCustomField(storyPointFieldId, currentColumn)
+        storyPointsDic.update(foundTasksDic)
+    # Differentiate between finished task cards that are labeled as approved and not approved
+    doneDic = {}
+    notDoneDic = {}
+    # Check if The current Task is labeled as approved or not and add it with the story points
+    # to the corresponding dictionary
+    for key, value in storyPointsDic.items():
+        if key in finishedTasks:
+            doneDic[key] = value
+            notDoneDic[key] = ""
+        else:
+            doneDic[key] = ""
+            notDoneDic[key] = value
+    # Create DataFrame hosting all Tasks and Story Points
+    storyPointsDataFrame = pd.DataFrame.from_dict(storyPointsDic, columns=["Story Points"], orient='index')
+    # Add Columns for Done and Unfinished Tasks
+    storyPointsDataFrame['Done'] = doneDic.values()
+    storyPointsDataFrame['Not Done'] = notDoneDic.values()
+    # Add Title to Index
+    storyPointsDataFrame.rename_axis("Task/StoryPoints", axis='index', inplace=True)
+
+    return storyPointsDataFrame
+
+
 # Write Dictionary locally to CSV File.
 def writeToCSV(taskDic, csvName):
     with open(csvName + '.csv', 'w') as output:
@@ -107,8 +173,6 @@ def writeToCSV(taskDic, csvName):
 
 
 def main():
-    # Dictionary holding the Tasks with the corresponding Story Points
-    storyPointsDic = {}
     # Id List of a Board and its relevant Columns
     boardId = 'Qt98t0nJ'
     sprintBacklogColumnId = '5e184a4bee152d1e2f34e434'
@@ -117,22 +181,14 @@ def main():
     # Id of Custom StoryPoint Field
     fieldStoryPointId = '5e5f7addad92130e3f2dbd60'
 
-    # Since we want to know how many story points are defined in the current sprint we need to add up the defined
-    # values of all 3 columns
-    ColumnIdList = [sprintBacklogColumnId, doingColumnId, reviewColumnId]
-
-    for i in ColumnIdList:
-        currentColumn = accessColumnCards(i)
-        foundTasksDic = filterCardsByCustomerField(fieldStoryPointId, currentColumn)
-        storyPointsDic.update(foundTasksDic)
-
-    # Write all Cards with Story Points to a CSV file
-    writeToCSV(storyPointsDic, "SprintStoryPoints")
-
     # Get all Cards from the Done Column and store them in a CSV File
     doneColumnId = '5e184a4648a22f2692e637ab'
+    taskColumns = [sprintBacklogColumnId, doingColumnId, reviewColumnId]
     columnDataFrame, _ = getColumnWithAllFields(boardId, doneColumnId)
     columnDataFrame.to_csv('FinishedTasks.csv', index=True)
+    # Generate Story Points Dataframe and csv for tracking
+    storyPointsDataFrame = generateStoryPointsDataFrame(boardId, taskColumns, reviewColumnId, fieldStoryPointId)
+    storyPointsDataFrame.to_csv("StoryPoints.csv", index=True)
 
 
 if __name__ == "__main__":