+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
-## 1.0.0
+## 1.1.0
 
 * Provide Methods for Trello Board Cards and Column Access via API
-
+* Provide Method which checks for all label definitions of a given board
+* Provide Method to filter a column for finished and unfinished task by checking for the 'approved' label
+